# docs/openapi.yaml
openapi: 3.0.3

info:
  title: go-api-starter
  version: 1.0.0
  description: |
    This contract is the source of truth. Handlers are generated from it, so an
    implementation that does not match will not compile.

    COMPATIBILITY RULES (v1):
      - Additive only. Removing a field, changing a type, or making a field
        required is breaking and CI rejects it (oasdiff).
      - The `code` value on Error is a public contract. Once a client uses it,
        it never changes.
      - Clients MUST tolerate enum values they do not recognise.

servers:
  - url: http://localhost:8080
    description: local

tags:
  - name: health
    description: Liveness and readiness
  - name: auth
    description: Registration, login and email verification
  - name: me
    description: The authenticated user
  - name: calculator
    description: Arithmetic

security:
  - bearerAuth: []

paths:
  /v1/auth/register:
    post:
      tags: [auth]
      operationId: registerUser
      summary: Create an account with an email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: |
            Account created. No session is issued: the email must be verified
            first, because auto-linking an SSO identity later is only safe when
            the local email was proven.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisteredUser"
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  /v1/auth/login:
    post:
      tags: [auth]
      operationId: loginUser
      summary: Exchange email and password for an access token
      security: []
      parameters:
        - $ref: "#/components/parameters/ClientPlatform"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: |
            Wrong credentials. The body is identical whether the email exists or
            the password is wrong, otherwise this endpoint becomes a user
            enumeration oracle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  /v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      summary: Rotate the refresh token and get a new access token
      description: |
        Every refresh mints a new token and spends the old one. Presenting a
        token that was already spent means two parties hold it, so the entire
        rotation chain is revoked and the session ends.
      security: []
      parameters:
        - $ref: "#/components/parameters/ClientPlatform"
        - $ref: "#/components/parameters/RefreshCookie"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: |
            Rotated. Web clients get the new refresh token in an httpOnly
            cookie and never in the body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: |
            Unknown, expired, revoked or already spent. The four are one answer
            so a client cannot probe for valid tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: End the session
      description: |
        Revokes the whole rotation chain, not just the token presented, so a
        copy taken earlier in the chain is dead too. Always succeeds: a client
        that wants out must never be told it is still signed in.
      security: []
      parameters:
        - $ref: "#/components/parameters/RefreshCookie"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Session ended

  /v1/auth/providers:
    get:
      tags: [auth]
      operationId: listProviders
      summary: Identity providers this server can actually serve
      description: |
        Both enabled in the database and configured on this server, so a client
        never renders a button that cannot work.
      security: []
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderInfo"

  /v1/auth/{provider}/start:
    get:
      tags: [auth]
      operationId: startProviderSignIn
      summary: Begin the browser sign-in flow
      description: |
        Records the state, nonce and PKCE verifier server side, then redirects
        to the provider. The state is deleted when the callback reads it, so a
        captured callback URL cannot be replayed.
      security: []
      parameters:
        - $ref: "#/components/parameters/ProviderCode"
        - name: redirect_to
          in: query
          required: false
          description: |
            Where to send the browser afterwards. Must be a path beginning with
            a single slash: accepting a full URL here would make this an open
            redirect.
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the provider
          headers:
            Location:
              schema:
                type: string
        "501": { $ref: "#/components/responses/NotImplemented" }

  /v1/auth/{provider}/callback:
    get:
      tags: [auth]
      operationId: completeProviderSignIn
      summary: Finish the browser sign-in flow
      description: |
        Sets the refresh cookie and redirects back to the app. The access token
        is deliberately not placed in the URL, where it would end up in browser
        history and referrer headers: the app calls /v1/auth/refresh to get one.
      security: []
      parameters:
        - $ref: "#/components/parameters/ProviderCode"
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect back to the app, with the refresh cookie set
          headers:
            Location:
              schema:
                type: string
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "501": { $ref: "#/components/responses/NotImplemented" }

  /v1/auth/{provider}/token:
    post:
      tags: [auth]
      operationId: signInWithProviderToken
      summary: Sign in with an ID token obtained by a native app
      description: |
        For native clients that got the token from the provider SDK. Verified
        by signature, issuer, audience allowlist and expiry. There is no nonce
        here because the client would be supplying both sides of it, which
        proves nothing; the browser flow is the stronger path.
      security: []
      parameters:
        - $ref: "#/components/parameters/ProviderCode"
        - $ref: "#/components/parameters/ClientPlatform"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderTokenRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "501": { $ref: "#/components/responses/NotImplemented" }

  /v1/me/identities:
    get:
      tags: [me]
      operationId: listIdentities
      summary: Providers linked to this account
      responses:
        "200":
          description: Linked identities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Identity"
        "401": { $ref: "#/components/responses/Unauthorized" }

  /v1/me/identities/{provider}:
    post:
      tags: [me]
      operationId: linkIdentity
      summary: Link a provider to this account
      description: |
        Needs no email checks: the caller has already proved they own this
        account. The unique index on provider and subject is what stops one
        provider account being attached to two users.
      parameters:
        - $ref: "#/components/parameters/ProviderCode"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderTokenRequest"
      responses:
        "201":
          description: Linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "501": { $ref: "#/components/responses/NotImplemented" }

  /v1/me/identities/{id}/unlink:
    post:
      tags: [me]
      operationId: unlinkIdentity
      summary: Remove a provider from this account
      description: |
        Refused when it is the only remaining way to sign in. Someone who
        signed up with a provider and never set a password would otherwise lock
        themselves out permanently.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Unlinked
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }

  /v1/auth/verify-email:
    post:
      tags: [auth]
      operationId: verifyEmail
      summary: Confirm an email address with a token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "204":
          description: Email verified
        "400": { $ref: "#/components/responses/BadRequest" }
        "410":
          description: Token expired or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/me:
    get:
      tags: [me]
      operationId: getMe
      summary: The authenticated user and its active login methods
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401": { $ref: "#/components/responses/Unauthorized" }

  /v1/calculations:
    post:
      tags: [calculator]
      operationId: calculate
      summary: Evaluate one arithmetic operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculationRequest"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculationResult"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  /healthz:
    get:
      tags: [health]
      operationId: getHealthz
      summary: Liveness probe
      description: Returns 200 while the process is alive. Touches no dependency.
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /readyz:
    get:
      tags: [health]
      operationId: getReadyz
      summary: Readiness probe
      description: Returns 200 only when every dependency is ready to serve traffic.
      responses:
        "200":
          description: Ready to serve traffic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"
        "503":
          description: A dependency is not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessStatus"

components:
  parameters:
    RefreshCookie:
      name: refresh_token
      in: cookie
      required: false
      description: |
        Where web clients carry the refresh token. Declared here so the cookie
        is part of the contract rather than an undocumented side channel.
      schema:
        type: string

    ProviderCode:
      name: provider
      in: path
      required: true
      description: Provider code, as listed by /v1/auth/providers.
      schema:
        type: string
        example: google

    ClientPlatform:
      name: X-Client-Platform
      in: header
      required: false
      description: |
        Decides how the refresh token is delivered. "web" gets it in an
        httpOnly cookie and never in the response body, so a script cannot read
        it. Anything else gets it in the body to put in secure storage.
      schema:
        type: string
        enum: [web, ios, android]

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflicts with existing state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: Well-formed but failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such thing, or not yours
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotImplemented:
      description: Not configured on this server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          # Deliberately not format: email. A bad address would then be
          # rejected while decoding, so the client would get a 400 with no
          # field detail for email and a 422 with detail for password. One
          # shape for every validation failure is worth more than the
          # generated type.
          type: string
          maxLength: 254
          example: someone@example.com
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 256
          description: |
            Minimum 12 characters. Length is the only rule: composition rules
            push people towards predictable substitutions.

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: someone@example.com
        password:
          type: string
          format: password

    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string

    RegisteredUser:
      type: object
      required: [id, email, email_verified]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean

    TokenPair:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Seconds until the access token expires.
          example: 900
        refresh_token:
          type: string
          description: |
            Absent for web clients: they receive it in an httpOnly cookie
            instead, which a script cannot read.

    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: |
            For clients that hold the token themselves. Web clients omit it
            because the cookie carries it.

    Me:
      type: object
      required: [id, email, email_verified, auth_methods]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        auth_methods:
          $ref: "#/components/schemas/AuthMethods"

    AuthMethods:
      type: object
      required: [password, identities]
      description: |
        What the client needs to decide between "Set password" and "Change
        password", and when to disable unlinking.
      properties:
        password:
          type: boolean
        identities:
          type: array
          items:
            $ref: "#/components/schemas/Identity"

    CalculationRequest:
      type: object
      required: [operation, a, b]
      properties:
        operation:
          type: string
          enum: [add, subtract, multiply, divide]
        a:
          $ref: "#/components/schemas/Operand"
        b:
          $ref: "#/components/schemas/Operand"

    CalculationResult:
      type: object
      required: [result]
      properties:
        result:
          type: string
          example: "13.5"

    Operand:
      type: string
      maxLength: 64
      description: |
        A decimal number, sent as a string on purpose. A JSON number is decoded
        into a float64 before any server code runs, and 0.1 + 0.2 is already
        wrong by that point. The length cap keeps a single request from asking
        for arbitrarily large arithmetic.
      example: "10.5"

    ProviderInfo:
      type: object
      required: [code, display_name]
      properties:
        code:
          type: string
          example: google
        display_name:
          type: string
          example: Google

    ProviderTokenRequest:
      type: object
      required: [id_token]
      properties:
        id_token:
          type: string

    Identity:
      type: object
      required: [id, provider]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
          example: google
        email:
          type: string

    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    ReadinessStatus:
      type: object
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessCheck"

    ReadinessCheck:
      type: object
      required: [name, ok]
      properties:
        name:
          type: string
          example: postgres
        ok:
          type: boolean
        error:
          type: string
          description: Set only when ok=false. Never carries internal detail.

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Stable code for clients to branch on. Never changes once released.
          example: ERR_AUTH_INVALID_CREDENTIALS
        message:
          type: string
          description: Human-readable. NOT a contract, may change at any time.
        details:
          type: array
          description: Per-field errors, for validation failures.
          items:
            $ref: "#/components/schemas/ErrorDetail"

    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string